refactor: split var and vars into different files like other structures
